Feat: [Be] 유저 관련 API 테스트 완료
유저 관련 API 테스트 완료

+++ backend/src/main/java/planeat/database/repository/UserCategoryInfoRepository.java
@@ -11,5 +11,6 @@ package planeat.database.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import planeat.database.entity.UserCategoryInfo;
 
-public interface UserCategoryInfoRepository extends JpaRepository<UserCategoryInfo, Long> {
+public interface UserCategoryInfoRepository extends JpaRepository<UserCategoryInfo, Integer> {
+
 }
